+++ CHANGELOG.md
@@ -28,6 +28,9 @@ Changelog
 
 **Next release**
 
+* Added support for [Dirty Unicorns](http://www.teamdirt.me/) (XPrivacy 3)
+* Added support for [Liquid Smooth](http://liquidsmooth.net/) (XPrivacy 3)
+* Added support for some CyanogenMod based ROMs, like Spirit ROM (XPrivacy 3)
 * Updated Dutch translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)